# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/admin/sync/football-data/leagues/{leagueExternalId}": {
      "post": {
        "operationId": "AdminSyncController_syncLeague",
        "parameters": [
          {
            "name": "leagueExternalId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/SyncJobResponseDto"
                    }
                  },
                  "required": [
                    "data"
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Tek ligi senkronla (async)",
        "tags": [
          "admin-sync"
        ]
      }
    },
...
    "/api/auth/register": {
      "post": {
        "operationId": "AuthController_register",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponseDto"
                }
              }
            }
          },
          "409": {
            "description": "E-posta/kullanıcı adı çakışması"
          }
        },
        "summary": "Kayıt ol",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Hatalı kimlik"
          }
        },
        "summary": "Giriş yap",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/forgot-password": {
      "post": {
        "operationId": "AuthController_forgotPassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ForgotPasswordDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "success",
                    "message"
                  ]
                }
              }
            }
          }
        },
        "summary": "Şifre sıfırlama isteği (her zaman 200)",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/reset-password": {
      "post": {
        "operationId": "AuthController_resetPassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResetPasswordDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "success",
                    "message"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Geçersiz/expired token veya zayıf parola"
          }
        },
        "summary": "Şifreyi sıfırla",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/refresh": {
      "post": {
        "operationId": "AuthController_refresh",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshTokenDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RefreshResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Geçersiz refresh token"
          }
        },
        "summary": "Access token yenile (refresh rotation)",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/logout": {
      "post": {
        "operationId": "AuthController_logout",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshTokenDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "success",
                    "message"
                  ]
                }
              }
            }
          }
        },
        "summary": "Çıkış (refresh token revoke)",
        "tags": [
          "auth"
        ]
      }
    },
...
    "/api/auth/google": {
      "post": {
        "operationId": "AuthController_google",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GoogleAuthDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponseDto"
                }
              }
            }
          },
          "503": {
            "description": "Google login etkin değil"
          }
        },
        "summary": "Google ile giriş",
        "tags": [
          "auth"
        ]
      }
    },
...
    "/api/admin/leagues": {
      "post": {
        "operationId": "AdminLeaguesController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LeagueWriteDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/LeagueCreatedResponseDto"
                    }
                  },
                  "required": [
                    "data"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Validation failed"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Lig olustur",
        "tags": [
          "admin-leagues"
        ]
      }
    },
    "/api/admin/leagues/{id}": {
      "put": {
        "operationId": "AdminLeaguesController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LeagueWriteDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed"
          },
          "404": {
            "description": "League not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Lig guncelle",
        "tags": [
          "admin-leagues"
        ]
      },
      "delete": {
        "operationId": "AdminLeaguesController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "League not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Lig sil",
        "tags": [
          "admin-leagues"
        ]
      }
    },
...
  },
  "info": {
    "title": "TransferPulse API",
    "description": "TransferPulse backend (.NET → NestJS migration)",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "SuccessResponseDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          }
        },
        "required": [
          "success"
        ]
      },
...
      "SyncJobResponseDto": {
        "type": "object",
        "properties": {
          "jobId": {
            "type": "string",
            "example": "42",
            "description": "BullMQ job ID"
          },
          "leagueExternalId": {
            "type": "number",
            "example": 7009,
            "description": "Sadece tek-lig sync isteğinde dolu"
          }
        },
        "required": [
          "jobId"
        ]
      },
...
      "RegisterDto": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "example": "alice",
            "minLength": 3,
            "maxLength": 50
          },
          "email": {
            "type": "string",
            "example": "alice@acme.com",
            "maxLength": 256
          },
          "password": {
            "type": "string",
            "example": "Secret123",
            "minLength": 8,
            "maxLength": 72
          },
          "nickname": {
            "type": "string",
            "example": "Alice",
            "maxLength": 50
          },
          "favouriteTeam": {
            "type": "string",
            "example": "Arsenal",
            "maxLength": 100
          }
        },
        "required": [
          "username",
          "email",
          "password",
          "nickname"
        ]
      },
      "UserResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "username": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "nickname": {
            "type": "string"
          },
          "profilePic": {
            "type": "string"
          },
          "isMailConfirm": {
            "type": "boolean"
          },
          "status": {
            "type": "string",
            "enum": [
              "Active",
              "Inactive",
              "Banned",
              "Suspended"
            ]
          },
          "favouriteTeam": {
            "type": "string"
          },
          "reputationScore": {
            "type": "number"
          },
          "role": {
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "username",
          "email",
          "nickname",
          "isMailConfirm",
          "status",
          "reputationScore",
          "role",
          "createdAt"
        ]
      },
      "AuthResponseDto": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string"
          },
          "refreshToken": {
            "type": "string"
          },
          "expiresAt": {
            "format": "date-time",
            "type": "string",
            "description": "Access token sona erme zamanı"
          },
          "user": {
            "$ref": "#/components/schemas/UserResponseDto"
          }
        },
        "required": [
          "accessToken",
          "refreshToken",
          "expiresAt",
          "user"
        ]
      },
      "LoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "alice@acme.com",
            "maxLength": 256
          },
          "password": {
            "type": "string",
            "example": "Secret123",
            "maxLength": 72
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "ForgotPasswordDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "alice@acme.com",
            "maxLength": 256
          }
        },
        "required": [
          "email"
        ]
      },
      "ResetPasswordDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "alice@acme.com",
            "maxLength": 256
          },
          "token": {
            "type": "string",
            "description": "E-posta ile gelen ham reset token"
          },
          "newPassword": {
            "type": "string",
            "example": "NewSecret123",
            "minLength": 8,
            "maxLength": 72
          }
        },
        "required": [
          "email",
          "token",
          "newPassword"
        ]
      },
      "RefreshTokenDto": {
        "type": "object",
        "properties": {
          "refreshToken": {
            "type": "string",
            "description": "Opaque refresh token (login/refresh yanıtından)"
          }
        },
        "required": [
          "refreshToken"
        ]
      },
      "RefreshResponseDto": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string",
            "description": "Yeni access token"
          },
          "refreshToken": {
            "type": "string",
            "description": "Yeni refresh token (rotation)"
          },
          "expiresAt": {
            "format": "date-time",
            "type": "string",
            "description": "Access token sona erme zamanı"
          }
        },
        "required": [
          "accessToken",
          "refreshToken",
          "expiresAt"
        ]
      },
      "GoogleAuthDto": {
        "type": "object",
        "properties": {
          "idToken": {
            "type": "string",
            "description": "Google ID token (client-side oturumdan)"
          }
        },
        "required": [
          "idToken"
        ]
      },
...
      "LeagueCreatedResponseDto": {
        "type": "object",
        "properties": {
          "leagueId": {
            "type": "string",
            "example": "550e8400-e29b-41d4-a716-446655440000"
          }
        },
        "required": [
          "leagueId"
        ]
      },
      "LeagueWriteDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 30
          },
          "country": {
            "type": "string",
            "maxLength": 30
          },
          "countryLogo": {
            "type": "string"
          },
          "leagueLogo": {
            "type": "string"
          },
          "leagueCode": {
            "type": "string",
            "maxLength": 10
          }
        },
        "required": [
          "name",
          "country",
          "countryLogo",
          "leagueLogo"
        ]
      },
...
    }
  }
}